Cleaned more garbage, reorganized things, added subtopics. Should be ready for meeting/use.
