--- constants/Colors.js
@@ -1,11 +1,0 @@
-const tintColor = '#2f95dc';
-
-export default {
-  tintColor,
-  errorBackground: 'red',
-  errorText: '#fff',
-  warningBackground: '#EAEB5E',
-  warningText: '#666804',
-  noticeBackground: tintColor,
-  noticeText: '#fff',
-};
