+++ screens/home/StylesHome.js
@@ -1,0 +1,7 @@
+import {StyleSheet} from 'react-native'
+
+const stylehome = StyleSheet.create({
+
+});
+
+export default stylehome;
