+++ screens/play/PlayScreen.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {ScrollView, Text, TouchableOpacity, View} from 'react-native';
-import styles from './StylesScreen'
+import styles from '../StylesScreen'
 
 export default class CreateScreen extends React.Component {
   render() { 
@@ -17,7 +17,3 @@ export default class CreateScreen extends React.Component {
     );
   }
 }
-
-CreateScreen.navigationOptions = {
-  header: null,
-};
