+++ screens/play/StylesPlay.js
@@ -1,0 +1,7 @@
+import {StyleSheet} from 'react-native'
+
+const styleplay = StyleSheet.create({
+
+});
+
+export default styleplay;
