+++ screens/settings/StylesSettings.js
@@ -1,0 +1,7 @@
+import {StyleSheet} from 'react-native'
+
+const stylesettings = StyleSheet.create({
+
+});
+
+export default stylesettings;
